refactor(TextEditor): fix TS errors (#2381)
